Separated lexer tests into new file
